Find links on a page

Find links on a page

Usage

ragnar_find_links(x, depth = 0L, children_only = TRUE, progress = TRUE)

Arguments

x

URL, HTML file path, or XML document. For Markdown, convert to HTML using commonmark::markdown_html() first.

depth

Integer specifying how many levels deep to crawl for links. When depth > 0, the function will follow child links (links with x as a prefix) and collect links from those pages as well.

children_only

Logical or string. If TRUE, returns only child links (those having x as a prefix). If FALSE, returns all links found on the page. Note that regardless of this setting, only child links are followed when depth > 0.

progress

Logical, draw a progress bar if depth > 0. A separate progress bar is drawn per recursion level.

Value

A character vector of links on the page.

Examples

ragnar_find_links("https://r4ds.hadley.nz/base-R.html")
#>  [1] "https://r4ds.hadley.nz/EDA.html"             
#>  [2] "https://r4ds.hadley.nz/arrow.html"           
#>  [3] "https://r4ds.hadley.nz/base-R.html"          
#>  [4] "https://r4ds.hadley.nz/communicate.html"     
#>  [5] "https://r4ds.hadley.nz/communication.html"   
#>  [6] "https://r4ds.hadley.nz/data-import.html"     
#>  [7] "https://r4ds.hadley.nz/data-tidy.html"       
#>  [8] "https://r4ds.hadley.nz/data-transform.html"  
#>  [9] "https://r4ds.hadley.nz/data-visualize.html"  
#> [10] "https://r4ds.hadley.nz/databases.html"       
#> [11] "https://r4ds.hadley.nz/datetimes.html"       
#> [12] "https://r4ds.hadley.nz/factors.html"         
#> [13] "https://r4ds.hadley.nz/functions.html"       
#> [14] "https://r4ds.hadley.nz/import.html"          
#> [15] "https://r4ds.hadley.nz/intro.html"           
#> [16] "https://r4ds.hadley.nz/iteration.html"       
#> [17] "https://r4ds.hadley.nz/joins.html"           
#> [18] "https://r4ds.hadley.nz/layers.html"          
#> [19] "https://r4ds.hadley.nz/logicals.html"        
#> [20] "https://r4ds.hadley.nz/missing-values.html"  
#> [21] "https://r4ds.hadley.nz/numbers.html"         
#> [22] "https://r4ds.hadley.nz/preface-2e.html"      
#> [23] "https://r4ds.hadley.nz/program.html"         
#> [24] "https://r4ds.hadley.nz/quarto-formats.html"  
#> [25] "https://r4ds.hadley.nz/quarto.html"          
#> [26] "https://r4ds.hadley.nz/rectangling.html"     
#> [27] "https://r4ds.hadley.nz/regexps.html"         
#> [28] "https://r4ds.hadley.nz/spreadsheets.html"    
#> [29] "https://r4ds.hadley.nz/strings.html"         
#> [30] "https://r4ds.hadley.nz/transform.html"       
#> [31] "https://r4ds.hadley.nz/visualize.html"       
#> [32] "https://r4ds.hadley.nz/webscraping.html"     
#> [33] "https://r4ds.hadley.nz/whole-game.html"      
#> [34] "https://r4ds.hadley.nz/workflow-basics.html" 
#> [35] "https://r4ds.hadley.nz/workflow-help.html"   
#> [36] "https://r4ds.hadley.nz/workflow-scripts.html"
#> [37] "https://r4ds.hadley.nz/workflow-style.html"  
ragnar_find_links("https://ellmer.tidyverse.org/")
#>  [1] "https://ellmer.tidyverse.org/LICENSE-text.html"                 
#>  [2] "https://ellmer.tidyverse.org/LICENSE.html"                      
#>  [3] "https://ellmer.tidyverse.org/articles/ellmer.html"              
#>  [4] "https://ellmer.tidyverse.org/articles/prompt-design.html"       
#>  [5] "https://ellmer.tidyverse.org/articles/streaming-async.html"     
#>  [6] "https://ellmer.tidyverse.org/articles/structured-data.html"     
#>  [7] "https://ellmer.tidyverse.org/articles/tool-calling.html"        
#>  [8] "https://ellmer.tidyverse.org/authors.html"                      
#>  [9] "https://ellmer.tidyverse.org/index.html"                        
#> [10] "https://ellmer.tidyverse.org/news/index.html"                   
#> [11] "https://ellmer.tidyverse.org/reference/chat_azure.html"         
#> [12] "https://ellmer.tidyverse.org/reference/chat_bedrock.html"       
#> [13] "https://ellmer.tidyverse.org/reference/chat_claude.html"        
#> [14] "https://ellmer.tidyverse.org/reference/chat_cortex_analyst.html"
#> [15] "https://ellmer.tidyverse.org/reference/chat_databricks.html"    
#> [16] "https://ellmer.tidyverse.org/reference/chat_deepseek.html"      
#> [17] "https://ellmer.tidyverse.org/reference/chat_gemini.html"        
#> [18] "https://ellmer.tidyverse.org/reference/chat_github.html"        
#> [19] "https://ellmer.tidyverse.org/reference/chat_groq.html"          
#> [20] "https://ellmer.tidyverse.org/reference/chat_ollama.html"        
#> [21] "https://ellmer.tidyverse.org/reference/chat_openai.html"        
#> [22] "https://ellmer.tidyverse.org/reference/chat_openrouter.html"    
#> [23] "https://ellmer.tidyverse.org/reference/chat_perplexity.html"    
#> [24] "https://ellmer.tidyverse.org/reference/chat_snowflake.html"     
#> [25] "https://ellmer.tidyverse.org/reference/chat_vllm.html"          
#> [26] "https://ellmer.tidyverse.org/reference/content_image_url.html"  
#> [27] "https://ellmer.tidyverse.org/reference/index.html"              
#> [28] "https://ellmer.tidyverse.org/reference/live_console.html"       
ragnar_find_links("https://ellmer.tidyverse.org/", depth = 2)
#> Following links... (recurse levels remaining: 1) ■■■■■■■■■■                    …
#> Following links... (recurse levels remaining: 1) ■■■■■■■■■■■■■■                …
#> Following links... (recurse levels remaining: 1) ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■…
#>  [1] "https://ellmer.tidyverse.org/LICENSE-text.html"                 
#>  [2] "https://ellmer.tidyverse.org/LICENSE.html"                      
#>  [3] "https://ellmer.tidyverse.org/articles/ellmer.html"              
#>  [4] "https://ellmer.tidyverse.org/articles/prompt-design.html"       
#>  [5] "https://ellmer.tidyverse.org/articles/streaming-async.html"     
#>  [6] "https://ellmer.tidyverse.org/articles/structured-data.html"     
#>  [7] "https://ellmer.tidyverse.org/articles/tool-calling.html"        
#>  [8] "https://ellmer.tidyverse.org/authors.html"                      
#>  [9] "https://ellmer.tidyverse.org/index.html"                        
#> [10] "https://ellmer.tidyverse.org/news/index.html"                   
#> [11] "https://ellmer.tidyverse.org/reference/Chat.html"               
#> [12] "https://ellmer.tidyverse.org/reference/Content.html"            
#> [13] "https://ellmer.tidyverse.org/reference/Provider.html"           
#> [14] "https://ellmer.tidyverse.org/reference/Turn.html"               
#> [15] "https://ellmer.tidyverse.org/reference/Type.html"               
#> [16] "https://ellmer.tidyverse.org/reference/chat_azure.html"         
#> [17] "https://ellmer.tidyverse.org/reference/chat_bedrock.html"       
#> [18] "https://ellmer.tidyverse.org/reference/chat_claude.html"        
#> [19] "https://ellmer.tidyverse.org/reference/chat_cortex.html"        
#> [20] "https://ellmer.tidyverse.org/reference/chat_cortex_analyst.html"
#> [21] "https://ellmer.tidyverse.org/reference/chat_databricks.html"    
#> [22] "https://ellmer.tidyverse.org/reference/chat_deepseek.html"      
#> [23] "https://ellmer.tidyverse.org/reference/chat_gemini.html"        
#> [24] "https://ellmer.tidyverse.org/reference/chat_github.html"        
#> [25] "https://ellmer.tidyverse.org/reference/chat_groq.html"          
#> [26] "https://ellmer.tidyverse.org/reference/chat_ollama.html"        
#> [27] "https://ellmer.tidyverse.org/reference/chat_openai.html"        
#> [28] "https://ellmer.tidyverse.org/reference/chat_openrouter.html"    
#> [29] "https://ellmer.tidyverse.org/reference/chat_perplexity.html"    
#> [30] "https://ellmer.tidyverse.org/reference/chat_snowflake.html"     
#> [31] "https://ellmer.tidyverse.org/reference/chat_vllm.html"          
#> [32] "https://ellmer.tidyverse.org/reference/content_image_url.html"  
#> [33] "https://ellmer.tidyverse.org/reference/content_pdf_file.html"   
#> [34] "https://ellmer.tidyverse.org/reference/contents_text.html"      
#> [35] "https://ellmer.tidyverse.org/reference/create_tool_def.html"    
#> [36] "https://ellmer.tidyverse.org/reference/gemini_upload.html"      
#> [37] "https://ellmer.tidyverse.org/reference/index.html"              
#> [38] "https://ellmer.tidyverse.org/reference/interpolate.html"        
#> [39] "https://ellmer.tidyverse.org/reference/live_console.html"       
#> [40] "https://ellmer.tidyverse.org/reference/token_usage.html"        
#> [41] "https://ellmer.tidyverse.org/reference/tool.html"               
#> [42] "https://ellmer.tidyverse.org/reference/type_boolean.html"       
ragnar_find_links("https://ellmer.tidyverse.org/", depth = 2, children_only = FALSE)
#>   [1] "http://tiktokenizer.vercel.app"                                                                             
#>   [2] "https://ai.google.dev/gemini-api/docs/prompting-intro"                                                      
#>   [3] "https://ai.google.dev/gemini-api/terms"                                                                     
#>   [4] "https://aistudio.google.com/app/apikey"                                                                     
#>   [5] "https://api-docs.deepseek.com/guides/function_calling"                                                      
#>   [6] "https://aws.amazon.com/bedrock"                                                                             
#>   [7] "https://aws.amazon.com/bedrock/claude"                                                                      
#>   [8] "https://aws.amazon.com/blogs/machine-learning/getting-started-with-cross-region-inference-in-amazon-bedrock"
#>   [9] "https://azure.microsoft.com/en-us/products/ai-services/openai-service"                                      
#>  [10] "https://bsky.app/profile/daviddiviny.com/post/3lb6kjaen4c2u"                                                
#>  [11] "https://cdn.openai.com/spec/model-spec-2024-05-08.html"                                                     
#>  [12] "https://chat.openai.com"                                                                                    
#>  [13] "https://cli.r-lib.org/reference/cli_abort.html"                                                             
#>  [14] "https://cloud.r-project.org/package=ellmer"                                                                 
#>  [15] "https://console.anthropic.com"                                                                              
#>  [16] "https://console.anthropic.com/account/keys"                                                                 
#>  [17] "https://coro.r-lib.org"                                                                                     
#>  [18] "https://coro.r-lib.org/articles/generator.html"                                                             
#>  [19] "https://coro.r-lib.org/reference/async.html"                                                                
#>  [20] "https://coro.r-lib.org/reference/async_generator.html"                                                      
#>  [21] "https://coro.r-lib.org/reference/collect.html"                                                              
#>  [22] "https://docs.anthropic.com/en/docs/build-with-claude/prompt-engineering/overview"                           
#>  [23] "https://docs.anthropic.com/en/docs/build-with-claude/prompt-engineering/prompt-generator"                   
#>  [24] "https://docs.anthropic.com/en/release-notes/system-prompts"                                                 
#>  [25] "https://docs.aws.amazon.com/bedrock/latest/APIReference/API_runtime_Converse.html"                          
#>  [26] "https://docs.databricks.com/en/dev-tools/auth/unified-auth.html"                                            
#>  [27] "https://docs.databricks.com/en/machine-learning/model-serving/score-foundation-models.html"                 
#>  [28] "https://docs.posit.co/connect/user/oauth-integrations"                                                      
#>  [29] "https://docs.posit.co/ide/server-pro/user/posit-workbench/managed-credentials/managed-credentials.html"     
#>  [30] "https://docs.ropensci.org/commonmark/reference/commonmark.html"                                             
#>  [31] "https://docs.snowflake.com/en/user-guide/admin-account-identifier"                                          
#>  [32] "https://docs.snowflake.com/en/user-guide/snowflake-cortex/cortex-analyst"                                   
#>  [33] "https://docs.snowflake.com/en/user-guide/snowflake-cortex/cortex-llm-rest-api"                              
#>  [34] "https://docs.vllm.ai/en/latest"                                                                             
#>  [35] "https://ellmer.tidyverse.org"                                                                               
#>  [36] "https://ellmer.tidyverse.org/LICENSE-text.html"                                                             
#>  [37] "https://ellmer.tidyverse.org/LICENSE.html"                                                                  
#>  [38] "https://ellmer.tidyverse.org/articles/ellmer.html"                                                          
#>  [39] "https://ellmer.tidyverse.org/articles/prompt-design.html"                                                   
#>  [40] "https://ellmer.tidyverse.org/articles/streaming-async.html"                                                 
#>  [41] "https://ellmer.tidyverse.org/articles/structured-data.html"                                                 
#>  [42] "https://ellmer.tidyverse.org/articles/tool-calling.html"                                                    
#>  [43] "https://ellmer.tidyverse.org/authors.html"                                                                  
#>  [44] "https://ellmer.tidyverse.org/index.html"                                                                    
#>  [45] "https://ellmer.tidyverse.org/news/index.html"                                                               
#>  [46] "https://ellmer.tidyverse.org/reference/Chat.html"                                                           
#>  [47] "https://ellmer.tidyverse.org/reference/Content.html"                                                        
#>  [48] "https://ellmer.tidyverse.org/reference/Provider.html"                                                       
#>  [49] "https://ellmer.tidyverse.org/reference/Turn.html"                                                           
#>  [50] "https://ellmer.tidyverse.org/reference/Type.html"                                                           
#>  [51] "https://ellmer.tidyverse.org/reference/chat_azure.html"                                                     
#>  [52] "https://ellmer.tidyverse.org/reference/chat_bedrock.html"                                                   
#>  [53] "https://ellmer.tidyverse.org/reference/chat_claude.html"                                                    
#>  [54] "https://ellmer.tidyverse.org/reference/chat_cortex.html"                                                    
#>  [55] "https://ellmer.tidyverse.org/reference/chat_cortex_analyst.html"                                            
#>  [56] "https://ellmer.tidyverse.org/reference/chat_databricks.html"                                                
#>  [57] "https://ellmer.tidyverse.org/reference/chat_deepseek.html"                                                  
#>  [58] "https://ellmer.tidyverse.org/reference/chat_gemini.html"                                                    
#>  [59] "https://ellmer.tidyverse.org/reference/chat_github.html"                                                    
#>  [60] "https://ellmer.tidyverse.org/reference/chat_groq.html"                                                      
#>  [61] "https://ellmer.tidyverse.org/reference/chat_ollama.html"                                                    
#>  [62] "https://ellmer.tidyverse.org/reference/chat_openai.html"                                                    
#>  [63] "https://ellmer.tidyverse.org/reference/chat_openrouter.html"                                                
#>  [64] "https://ellmer.tidyverse.org/reference/chat_perplexity.html"                                                
#>  [65] "https://ellmer.tidyverse.org/reference/chat_snowflake.html"                                                 
#>  [66] "https://ellmer.tidyverse.org/reference/chat_vllm.html"                                                      
#>  [67] "https://ellmer.tidyverse.org/reference/content_image_url.html"                                              
#>  [68] "https://ellmer.tidyverse.org/reference/content_pdf_file.html"                                               
#>  [69] "https://ellmer.tidyverse.org/reference/contents_text.html"                                                  
#>  [70] "https://ellmer.tidyverse.org/reference/create_tool_def.html"                                                
#>  [71] "https://ellmer.tidyverse.org/reference/gemini_upload.html"                                                  
#>  [72] "https://ellmer.tidyverse.org/reference/index.html"                                                          
#>  [73] "https://ellmer.tidyverse.org/reference/interpolate.html"                                                    
#>  [74] "https://ellmer.tidyverse.org/reference/live_console.html"                                                   
#>  [75] "https://ellmer.tidyverse.org/reference/token_usage.html"                                                    
#>  [76] "https://ellmer.tidyverse.org/reference/tool.html"                                                           
#>  [77] "https://ellmer.tidyverse.org/reference/type_boolean.html"                                                   
#>  [78] "https://gh.r-lib.org"                                                                                       
#>  [79] "https://gist.github.com/dannguyen/faaa56cebf30ad51108a9fe4f8db36d8"                                         
#>  [80] "https://gist.github.com/hadley/7688b4dd1e5e97b800c6d7d79e437b48"                                            
#>  [81] "https://github.com/andrie"                                                                                  
#>  [82] "https://github.com/anthropics/anthropic-cookbook/blob/main/tool_use/extracting_structured_json.ipynb"       
#>  [83] "https://github.com/atheriel"                                                                                
#>  [84] "https://github.com/billsanto"                                                                               
#>  [85] "https://github.com/cpsievert/aidea"                                                                         
#>  [86] "https://github.com/cpsievert/aidea/blob/main/inst/app/prompt.md"                                            
#>  [87] "https://github.com/jcheng5/elmer-assistant"                                                                 
#>  [88] "https://github.com/jcheng5/py-sidebot/blob/main/prompt.md"                                                  
#>  [89] "https://github.com/jcheng5/shinychat"                                                                       
#>  [90] "https://github.com/marketplace/models"                                                                      
#>  [91] "https://github.com/posit-dev/chatlas"                                                                       
#>  [92] "https://github.com/posit-dev/shiny-assistant/blob/main/shinyapp/app_prompt.md"                              
#>  [93] "https://github.com/posit-dev/shiny-assistant/blob/main/shinyapp/app_prompt_python.md"                       
#>  [94] "https://github.com/posit-dev/shiny-assistant/blob/main/shinyapp/app_prompt_r.md"                            
#>  [95] "https://github.com/simonpcouch/pal/tree/main/inst/prompts"                                                  
#>  [96] "https://github.com/tidyverse/ellmer"                                                                        
#>  [97] "https://github.com/tidyverse/ellmer/blob/main/DESCRIPTION"                                                  
#>  [98] "https://github.com/tidyverse/ellmer/blob/main/LICENSE.md"                                                   
#>  [99] "https://github.com/tidyverse/ellmer/blob/main/NEWS.md"                                                      
#> [100] "https://github.com/tidyverse/ellmer/blob/main/R/chat.R"                                                     
#> [101] "https://github.com/tidyverse/ellmer/blob/main/R/content-image.R"                                            
#> [102] "https://github.com/tidyverse/ellmer/blob/main/R/content-pdf.R"                                              
#> [103] "https://github.com/tidyverse/ellmer/blob/main/R/content.R"                                                  
#> [104] "https://github.com/tidyverse/ellmer/blob/main/R/interpolate.R"                                              
#> [105] "https://github.com/tidyverse/ellmer/blob/main/R/provider-azure.R"                                           
#> [106] "https://github.com/tidyverse/ellmer/blob/main/R/provider-bedrock.R"                                         
#> [107] "https://github.com/tidyverse/ellmer/blob/main/R/provider-claude.R"                                          
#> [108] "https://github.com/tidyverse/ellmer/blob/main/R/provider-cortex.R"                                          
#> [109] "https://github.com/tidyverse/ellmer/blob/main/R/provider-databricks.R"                                      
#> [110] "https://github.com/tidyverse/ellmer/blob/main/R/provider-deepseek.R"                                        
#> [111] "https://github.com/tidyverse/ellmer/blob/main/R/provider-gemini-upload.R"                                   
#> [112] "https://github.com/tidyverse/ellmer/blob/main/R/provider-gemini.R"                                          
#> [113] "https://github.com/tidyverse/ellmer/blob/main/R/provider-github.R"                                          
#> [114] "https://github.com/tidyverse/ellmer/blob/main/R/provider-groq.R"                                            
#> [115] "https://github.com/tidyverse/ellmer/blob/main/R/provider-ollama.R"                                          
#> [116] "https://github.com/tidyverse/ellmer/blob/main/R/provider-openai.R"                                          
#> [117] "https://github.com/tidyverse/ellmer/blob/main/R/provider-openrouter.R"                                      
#> [118] "https://github.com/tidyverse/ellmer/blob/main/R/provider-perplexity.R"                                      
#> [119] "https://github.com/tidyverse/ellmer/blob/main/R/provider-snowflake.R"                                       
#> [120] "https://github.com/tidyverse/ellmer/blob/main/R/provider-vllm.R"                                            
#> [121] "https://github.com/tidyverse/ellmer/blob/main/R/provider.R"                                                 
#> [122] "https://github.com/tidyverse/ellmer/blob/main/R/shiny.R"                                                    
#> [123] "https://github.com/tidyverse/ellmer/blob/main/R/tokens.R"                                                   
#> [124] "https://github.com/tidyverse/ellmer/blob/main/R/tools-def-auto.R"                                           
#> [125] "https://github.com/tidyverse/ellmer/blob/main/R/tools-def.R"                                                
#> [126] "https://github.com/tidyverse/ellmer/blob/main/R/turns.R"                                                    
#> [127] "https://github.com/tidyverse/ellmer/blob/main/R/types.R"                                                    
#> [128] "https://github.com/tidyverse/ellmer/blob/main/vignettes/ellmer.Rmd"                                         
#> [129] "https://github.com/tidyverse/ellmer/blob/main/vignettes/prompt-design.Rmd"                                  
#> [130] "https://github.com/tidyverse/ellmer/blob/main/vignettes/streaming-async.Rmd"                                
#> [131] "https://github.com/tidyverse/ellmer/blob/main/vignettes/structured-data.Rmd"                                
#> [132] "https://github.com/tidyverse/ellmer/blob/main/vignettes/tool-calling.Rmd"                                   
#> [133] "https://github.com/tidyverse/ellmer/issues"                                                                 
#> [134] "https://github.com/tidyverse/ellmer/issues/110"                                                             
#> [135] "https://github.com/tidyverse/ellmer/issues/115"                                                             
#> [136] "https://github.com/tidyverse/ellmer/issues/117"                                                             
#> [137] "https://github.com/tidyverse/ellmer/issues/140"                                                             
#> [138] "https://github.com/tidyverse/ellmer/issues/143"                                                             
#> [139] "https://github.com/tidyverse/ellmer/issues/152"                                                             
#> [140] "https://github.com/tidyverse/ellmer/issues/157"                                                             
#> [141] "https://github.com/tidyverse/ellmer/issues/212"                                                             
#> [142] "https://github.com/tidyverse/ellmer/issues/213"                                                             
#> [143] "https://github.com/tidyverse/ellmer/issues/242"                                                             
#> [144] "https://github.com/tidyverse/ellmer/issues/246"                                                             
#> [145] "https://github.com/tidyverse/ellmer/issues/248"                                                             
#> [146] "https://github.com/tidyverse/ellmer/issues/252"                                                             
#> [147] "https://github.com/tidyverse/ellmer/issues/254"                                                             
#> [148] "https://github.com/tidyverse/ellmer/issues/257"                                                             
#> [149] "https://github.com/tidyverse/ellmer/issues/258"                                                             
#> [150] "https://github.com/tidyverse/ellmer/issues/261"                                                             
#> [151] "https://github.com/tidyverse/ellmer/issues/263"                                                             
#> [152] "https://github.com/tidyverse/ellmer/issues/265"                                                             
#> [153] "https://github.com/tidyverse/ellmer/issues/269"                                                             
#> [154] "https://github.com/tidyverse/ellmer/issues/271"                                                             
#> [155] "https://github.com/tidyverse/ellmer/issues/273"                                                             
#> [156] "https://github.com/tidyverse/ellmer/issues/275"                                                             
#> [157] "https://github.com/tidyverse/ellmer/issues/285"                                                             
#> [158] "https://github.com/tidyverse/ellmer/issues/295"                                                             
#> [159] "https://github.com/tidyverse/ellmer/issues/299"                                                             
#> [160] "https://github.com/tidyverse/ellmer/issues/300"                                                             
#> [161] "https://github.com/tidyverse/ellmer/issues/31"                                                              
#> [162] "https://github.com/tidyverse/ellmer/issues/310"                                                             
#> [163] "https://github.com/tidyverse/ellmer/issues/317"                                                             
#> [164] "https://github.com/tidyverse/ellmer/issues/32"                                                              
#> [165] "https://github.com/tidyverse/ellmer/issues/320"                                                             
#> [166] "https://github.com/tidyverse/ellmer/issues/50"                                                              
#> [167] "https://github.com/tidyverse/ellmer/issues/56"                                                              
#> [168] "https://github.com/walkerke"                                                                                
#> [169] "https://glue.tidyverse.org"                                                                                 
#> [170] "https://glue.tidyverse.org/articles/wrappers.html"                                                          
#> [171] "https://groq.com"                                                                                           
#> [172] "https://hadley.nz"                                                                                          
#> [173] "https://hauselin.github.io/ollama-r"                                                                        
#> [174] "https://httr2.r-lib.org/reference/req_headers.html"                                                         
#> [175] "https://httr2.r-lib.org/reference/with_verbosity.html"                                                      
#> [176] "https://json-schema.org"                                                                                    
#> [177] "https://lifecycle.r-lib.org/articles/stages.html"                                                           
#> [178] "https://ollama.com"                                                                                         
#> [179] "https://openai.com"                                                                                         
#> [180] "https://openrouter.ai"                                                                                      
#> [181] "https://openrouter.ai/models"                                                                               
#> [182] "https://opensource.org/licenses/mit-license.php"                                                            
#> [183] "https://orcid.org/0000-0003-4757-117X"                                                                      
#> [184] "https://pkgdown.r-lib.org"                                                                                  
#> [185] "https://platform.deepseek.com"                                                                              
#> [186] "https://platform.openai.com"                                                                                
#> [187] "https://platform.openai.com/account/api-keys"                                                               
#> [188] "https://platform.openai.com/docs/guides/prompt-engineering"                                                 
#> [189] "https://platform.openai.com/docs/guides/vision/low-or-high-fidelity-image-understanding"                    
#> [190] "https://posit-dev.github.io/brand-yml/articles/llm-brand-yml-prompt"                                        
#> [191] "https://r-lib.github.io/later/reference/later_fd.html"                                                      
#> [192] "https://r6.r-lib.org"                                                                                       
#> [193] "https://rdrr.io/r/base/AsIs.html"                                                                           
#> [194] "https://rdrr.io/r/base/Paren.html"                                                                          
#> [195] "https://rdrr.io/r/base/Sys.time.html"                                                                       
#> [196] "https://rdrr.io/r/base/c.html"                                                                              
#> [197] "https://rdrr.io/r/base/cat.html"                                                                            
#> [198] "https://rdrr.io/r/base/character.html"                                                                      
#> [199] "https://rdrr.io/r/base/chartr.html"                                                                         
#> [200] "https://rdrr.io/r/base/do.call.html"                                                                        
#> [201] "https://rdrr.io/r/base/format.html"                                                                         
#> [202] "https://rdrr.io/r/base/interactive.html"                                                                    
#> [203] "https://rdrr.io/r/base/lapply.html"                                                                         
#> [204] "https://rdrr.io/r/base/library.html"                                                                        
#> [205] "https://rdrr.io/r/base/list.html"                                                                           
#> [206] "https://rdrr.io/r/base/print.html"                                                                          
#> [207] "https://rdrr.io/r/base/readLines.html"                                                                      
#> [208] "https://rdrr.io/r/base/stop.html"                                                                           
#> [209] "https://rdrr.io/r/base/sys.parent.html"                                                                     
#> [210] "https://rdrr.io/r/base/system.file.html"                                                                    
#> [211] "https://rdrr.io/r/graphics/plot.default.html"                                                               
#> [212] "https://rdrr.io/r/stats/Normal.html"                                                                        
#> [213] "https://rdrr.io/r/utils/install.packages.html"                                                              
#> [214] "https://rdrr.io/r/utils/str.html"                                                                           
#> [215] "https://rlang.r-lib.org/reference/is_installed.html"                                                        
#> [216] "https://roxygen2.r-lib.org"                                                                                 
#> [217] "https://rstudio.github.io/promises"                                                                         
#> [218] "https://rstudio.github.io/promises/reference/pipes.html"                                                    
#> [219] "https://shiny.posit.co/blog/posts/shiny-assistant"                                                          
#> [220] "https://simonpcouch.github.io/pal"                                                                          
#> [221] "https://usethis.r-lib.org/articles/git-credentials.html"                                                    
#> [222] "https://www.anthropic.com"                                                                                  
#> [223] "https://www.anthropic.com/claude"                                                                           
#> [224] "https://www.oneusefulthing.org/p/getting-started-with-ai-good-enough"                                       
#> [225] "https://www.paws-r-sdk.com"                                                                                 
#> [226] "https://www.perplexity.ai"                                                                                  
#> [227] "https://www.posit.co"                                                                                       
#> [228] "https://www.youtube.com/watch?v=sYliwvml9Es"                                                                
#> [229] "https://www.youtube.com/watch?v=zduSFxRajkE"                                                                
ragnar_find_links(
  "https://github.com/Snowflake-Labs/sfquickstarts/tree/master/site/sfguides/src/build_a_custom_model_for_anomaly_detection",
  children_only = "https://github.com/Snowflake-Labs/sfquickstarts",
  depth = 1
)
#>  [1] "https://github.com/Snowflake-Labs/sfquickstarts"                                                                                                                       
#>  [2] "https://github.com/Snowflake-Labs/sfquickstarts/actions"                                                                                                               
#>  [3] "https://github.com/Snowflake-Labs/sfquickstarts/blob/master/site/sfguides/src/build_a_custom_model_for_anomaly_detection/assets/external_access.gif"                   
#>  [4] "https://github.com/Snowflake-Labs/sfquickstarts/blob/master/site/sfguides/src/build_a_custom_model_for_anomaly_detection/build_a_custom_model_for_anomaly_detection.md"
#>  [5] "https://github.com/Snowflake-Labs/sfquickstarts/commits/master/site/sfguides/src/build_a_custom_model_for_anomaly_detection"                                           
#>  [6] "https://github.com/Snowflake-Labs/sfquickstarts/commits/master/site/sfguides/src/build_a_custom_model_for_anomaly_detection/assets"                                    
#>  [7] "https://github.com/Snowflake-Labs/sfquickstarts/issues"                                                                                                                
#>  [8] "https://github.com/Snowflake-Labs/sfquickstarts/projects"                                                                                                              
#>  [9] "https://github.com/Snowflake-Labs/sfquickstarts/pulls"                                                                                                                 
#> [10] "https://github.com/Snowflake-Labs/sfquickstarts/pulse"                                                                                                                 
#> [11] "https://github.com/Snowflake-Labs/sfquickstarts/security"                                                                                                              
#> [12] "https://github.com/Snowflake-Labs/sfquickstarts/tree/master"                                                                                                           
#> [13] "https://github.com/Snowflake-Labs/sfquickstarts/tree/master/site"                                                                                                      
#> [14] "https://github.com/Snowflake-Labs/sfquickstarts/tree/master/site/sfguides"                                                                                             
#> [15] "https://github.com/Snowflake-Labs/sfquickstarts/tree/master/site/sfguides/src"                                                                                         
#> [16] "https://github.com/Snowflake-Labs/sfquickstarts/tree/master/site/sfguides/src/build_a_custom_model_for_anomaly_detection"                                              
#> [17] "https://github.com/Snowflake-Labs/sfquickstarts/tree/master/site/sfguides/src/build_a_custom_model_for_anomaly_detection/assets"