Find links on a page

Find links on a page

Usage

ragnar_find_links(
  x,
  depth = 0L,
  children_only = FALSE,
  progress = TRUE,
  ...,
  url_filter = identity,
  validate = FALSE
)

Arguments

x

URL, HTML file path, or XML document. For Markdown, convert to HTML using commonmark::markdown_html() first.

depth

Integer specifying how many levels deep to crawl for links. When depth > 0, the function will follow child links (links with x as a prefix) and collect links from those pages as well.

children_only

Logical or string. If TRUE, returns only child links (those having x as a prefix). If FALSE, returns all links found on the page. Note that regardless of this setting, only child links are followed when depth > 0.

progress

Logical, draw a progress bar if depth > 0.

...

Currently unused. Must be empty.

url_filter

A function that takes a character vector of URL's and may subset them to return a smaller list. This can be useful for filtering out URL's by rules different than children_only which only checks the prefix.

validate

Default is FALSE. If TRUE sends a HEAD request for each link and removes those that are not accessible. Requests are sent in parallel using httr2::req_perform_parallel().

Value

A character vector of links on the page.

Examples

# \dontrun{
ragnar_find_links("https://r4ds.hadley.nz/base-R.html")
#>  [1] "https://adv-r.hadley.nz/subsetting.html"                        
#>  [2] "https://dplyr.tidyverse.org/reference/across.html"              
#>  [3] "https://dplyr.tidyverse.org/reference/arrange.html"             
#>  [4] "https://dplyr.tidyverse.org/reference/filter.html"              
#>  [5] "https://dplyr.tidyverse.org/reference/group_by.html"            
#>  [6] "https://dplyr.tidyverse.org/reference/mutate.html"              
#>  [7] "https://dplyr.tidyverse.org/reference/pull.html"                
#>  [8] "https://dplyr.tidyverse.org/reference/relocate.html"            
#>  [9] "https://dplyr.tidyverse.org/reference/select.html"              
#> [10] "https://dplyr.tidyverse.org/reference/summarise.html"           
#> [11] "https://gist.github.com/hadley/1986a273e384fb2d4d752c18ed71bedf"
#> [12] "https://gist.github.com/hadley/c430501804349d382ce90754936ab8ec"
#> [13] "https://github.com/hadley/r4ds"                                 
#> [14] "https://github.com/hadley/r4ds/edit/main/base-R.qmd"            
#> [15] "https://github.com/hadley/r4ds/issues/new"                      
#> [16] "https://purrr.tidyverse.org/reference/map.html"                 
#> [17] "https://quarto.org"                                             
#> [18] "https://r4ds.hadley.nz/EDA.html"                                
#> [19] "https://r4ds.hadley.nz/arrow.html"                              
#> [20] "https://r4ds.hadley.nz/base-R.html"                             
#> [21] "https://r4ds.hadley.nz/communicate.html"                        
#> [22] "https://r4ds.hadley.nz/communication.html"                      
#> [23] "https://r4ds.hadley.nz/data-import.html"                        
#> [24] "https://r4ds.hadley.nz/data-tidy.html"                          
#> [25] "https://r4ds.hadley.nz/data-transform.html"                     
#> [26] "https://r4ds.hadley.nz/data-visualize.html"                     
#> [27] "https://r4ds.hadley.nz/databases.html"                          
#> [28] "https://r4ds.hadley.nz/datetimes.html"                          
#> [29] "https://r4ds.hadley.nz/factors.html"                            
#> [30] "https://r4ds.hadley.nz/functions.html"                          
#> [31] "https://r4ds.hadley.nz/import.html"                             
#> [32] "https://r4ds.hadley.nz/intro.html"                              
#> [33] "https://r4ds.hadley.nz/iteration.html"                          
#> [34] "https://r4ds.hadley.nz/joins.html"                              
#> [35] "https://r4ds.hadley.nz/layers.html"                             
#> [36] "https://r4ds.hadley.nz/logicals.html"                           
#> [37] "https://r4ds.hadley.nz/missing-values.html"                     
#> [38] "https://r4ds.hadley.nz/numbers.html"                            
#> [39] "https://r4ds.hadley.nz/preface-2e.html"                         
#> [40] "https://r4ds.hadley.nz/program.html"                            
#> [41] "https://r4ds.hadley.nz/quarto-formats.html"                     
#> [42] "https://r4ds.hadley.nz/quarto.html"                             
#> [43] "https://r4ds.hadley.nz/rectangling.html"                        
#> [44] "https://r4ds.hadley.nz/regexps.html"                            
#> [45] "https://r4ds.hadley.nz/spreadsheets.html"                       
#> [46] "https://r4ds.hadley.nz/strings.html"                            
#> [47] "https://r4ds.hadley.nz/transform.html"                          
#> [48] "https://r4ds.hadley.nz/visualize.html"                          
#> [49] "https://r4ds.hadley.nz/webscraping.html"                        
#> [50] "https://r4ds.hadley.nz/whole-game.html"                         
#> [51] "https://r4ds.hadley.nz/workflow-basics.html"                    
#> [52] "https://r4ds.hadley.nz/workflow-help.html"                      
#> [53] "https://r4ds.hadley.nz/workflow-scripts.html"                   
#> [54] "https://r4ds.hadley.nz/workflow-style.html"                     
#> [55] "https://rdrr.io/r/base/Arithmetic.html"                         
#> [56] "https://rdrr.io/r/base/Extremes.html"                           
#> [57] "https://rdrr.io/r/base/NA.html"                                 
#> [58] "https://rdrr.io/r/base/apply.html"                              
#> [59] "https://rdrr.io/r/base/c.html"                                  
#> [60] "https://rdrr.io/r/base/cbind.html"                              
#> [61] "https://rdrr.io/r/base/data.frame.html"                         
#> [62] "https://rdrr.io/r/base/do.call.html"                            
#> [63] "https://rdrr.io/r/base/lapply.html"                             
#> [64] "https://rdrr.io/r/base/length.html"                             
#> [65] "https://rdrr.io/r/base/levels.html"                             
#> [66] "https://rdrr.io/r/base/library.html"                            
#> [67] "https://rdrr.io/r/base/list.files.html"                         
#> [68] "https://rdrr.io/r/base/list.html"                               
#> [69] "https://rdrr.io/r/base/logical.html"                            
#> [70] "https://rdrr.io/r/base/mean.html"                               
#> [71] "https://rdrr.io/r/base/numeric.html"                            
#> [72] "https://rdrr.io/r/base/order.html"                              
#> [73] "https://rdrr.io/r/base/seq.html"                                
#> [74] "https://rdrr.io/r/base/subset.html"                             
#> [75] "https://rdrr.io/r/base/sum.html"                                
#> [76] "https://rdrr.io/r/base/tapply.html"                             
#> [77] "https://rdrr.io/r/base/transform.html"                          
#> [78] "https://rdrr.io/r/base/vector.html"                             
#> [79] "https://rdrr.io/r/base/which.html"                              
#> [80] "https://rdrr.io/r/base/with.html"                               
#> [81] "https://rdrr.io/r/graphics/hist.html"                           
#> [82] "https://rdrr.io/r/graphics/plot.default.html"                   
#> [83] "https://rdrr.io/r/stats/Uniform.html"                           
#> [84] "https://rdrr.io/r/utils/str.html"                               
#> [85] "https://readxl.tidyverse.org/reference/read_excel.html"         
#> [86] "https://tibble.tidyverse.org/reference/tibble.html"             
#> [87] "https://tidyselect.r-lib.org/reference/starts_with.html"        
#> [88] "https://tidyverse.tidyverse.org"                                
ragnar_find_links("https://ellmer.tidyverse.org/")
#>  [1] "http://schloerke.com"                                                                                  
#>  [2] "https://ai.google.dev/gemini-api/terms"                                                                
#>  [3] "https://cloud.r-project.org/package=ellmer"                                                            
#>  [4] "https://docs.posit.co/connect/user/oauth-integrations"                                                 
#>  [5] "https://docs.posit.co/ide/server-pro/user/posit-workbench/managed-credentials/managed-credentials.html"
#>  [6] "https://ellmer.tidyverse.org"                                                                          
#>  [7] "https://ellmer.tidyverse.org/LICENSE-text.html"                                                        
#>  [8] "https://ellmer.tidyverse.org/LICENSE.html"                                                             
#>  [9] "https://ellmer.tidyverse.org/articles/ellmer.html"                                                     
#> [10] "https://ellmer.tidyverse.org/articles/programming.html"                                                
#> [11] "https://ellmer.tidyverse.org/articles/prompt-design.html"                                              
#> [12] "https://ellmer.tidyverse.org/articles/streaming-async.html"                                            
#> [13] "https://ellmer.tidyverse.org/articles/structured-data.html"                                            
#> [14] "https://ellmer.tidyverse.org/articles/tool-calling.html"                                               
#> [15] "https://ellmer.tidyverse.org/authors.html"                                                             
#> [16] "https://ellmer.tidyverse.org/index.html"                                                               
#> [17] "https://ellmer.tidyverse.org/news/index.html"                                                          
#> [18] "https://ellmer.tidyverse.org/reference/chat-any.html"                                                  
#> [19] "https://ellmer.tidyverse.org/reference/chat_anthropic.html"                                            
#> [20] "https://ellmer.tidyverse.org/reference/chat_aws_bedrock.html"                                          
#> [21] "https://ellmer.tidyverse.org/reference/chat_azure_openai.html"                                         
#> [22] "https://ellmer.tidyverse.org/reference/chat_cloudflare.html"                                           
#> [23] "https://ellmer.tidyverse.org/reference/chat_cortex_analyst.html"                                       
#> [24] "https://ellmer.tidyverse.org/reference/chat_databricks.html"                                           
#> [25] "https://ellmer.tidyverse.org/reference/chat_deepseek.html"                                             
#> [26] "https://ellmer.tidyverse.org/reference/chat_github.html"                                               
#> [27] "https://ellmer.tidyverse.org/reference/chat_google_gemini.html"                                        
#> [28] "https://ellmer.tidyverse.org/reference/chat_groq.html"                                                 
#> [29] "https://ellmer.tidyverse.org/reference/chat_huggingface.html"                                          
#> [30] "https://ellmer.tidyverse.org/reference/chat_mistral.html"                                              
#> [31] "https://ellmer.tidyverse.org/reference/chat_ollama.html"                                               
#> [32] "https://ellmer.tidyverse.org/reference/chat_openai.html"                                               
#> [33] "https://ellmer.tidyverse.org/reference/chat_openrouter.html"                                           
#> [34] "https://ellmer.tidyverse.org/reference/chat_perplexity.html"                                           
#> [35] "https://ellmer.tidyverse.org/reference/chat_snowflake.html"                                            
#> [36] "https://ellmer.tidyverse.org/reference/chat_vllm.html"                                                 
#> [37] "https://ellmer.tidyverse.org/reference/content_image_url.html"                                         
#> [38] "https://ellmer.tidyverse.org/reference/index.html"                                                     
#> [39] "https://ellmer.tidyverse.org/reference/live_console.html"                                              
#> [40] "https://garrickadenbuie.com"                                                                           
#> [41] "https://github.com/posit-dev/chatlas"                                                                  
#> [42] "https://github.com/tidyverse/ellmer"                                                                   
#> [43] "https://github.com/tidyverse/ellmer/issues"                                                            
#> [44] "https://hadley.nz"                                                                                     
#> [45] "https://ollama.com"                                                                                    
#> [46] "https://opensource.org/licenses/mit-license.php"                                                       
#> [47] "https://orcid.org/0000-0001-9986-114X"                                                                 
#> [48] "https://orcid.org/0000-0002-7111-0077"                                                                 
#> [49] "https://orcid.org/0000-0003-4757-117X"                                                                 
#> [50] "https://pkgdown.r-lib.org"                                                                             
#> [51] "https://posit-dev.github.io/mcptools"                                                                  
#> [52] "https://posit-dev.github.io/shinychat"                                                                 
#> [53] "https://posit.co/blog/announcing-ellmer"                                                               
#> [54] "https://r6.r-lib.org"                                                                                  
#> [55] "https://ragnar.tidyverse.org"                                                                          
#> [56] "https://rdrr.io/r/base/library.html"                                                                   
#> [57] "https://rdrr.io/r/utils/install.packages.html"                                                         
#> [58] "https://rdrr.io/r/utils/str.html"                                                                      
#> [59] "https://ror.org/03wc8by49"                                                                             
#> [60] "https://vitals.tidyverse.org"                                                                          
#> [61] "https://www.posit.co"                                                                                  
#> [62] "https://www.tidyverse.org/blog/2025/05/ellmer-0-2-0"                                                   
#> [63] "https://www.tidyverse.org/blog/2025/07/ellmer-0-3-0"                                                   
ragnar_find_links(
  paste0("https://github.com/Snowflake-Labs/sfquickstarts/",
         "tree/master/site/sfguides/src/build_a_custom_model_for_anomaly_detection"),
  children_only = "https://github.com/Snowflake-Labs/sfquickstarts",
  depth = 1
)
#> ⠙ Finding links: 4 | On queue: 3 | Current depth: 1 | [5s]
#> ⠹ Finding links: 8 | On queue: 0 | Current depth: 1 | [7s]
#>   [1] "https://github.com/Snowflake-Labs/sfquickstarts"                                                                          
#>   [2] "https://github.com/Snowflake-Labs/sfquickstarts/actions"                                                                  
#>   [3] "https://github.com/Snowflake-Labs/sfquickstarts/actions/caches"                                                           
#>   [4] "https://github.com/Snowflake-Labs/sfquickstarts/actions/runs/18788363058"                                                 
#>   [5] "https://github.com/Snowflake-Labs/sfquickstarts/actions/runs/18788363058/workflow"                                        
#>   [6] "https://github.com/Snowflake-Labs/sfquickstarts/actions/runs/18788729669"                                                 
#>   [7] "https://github.com/Snowflake-Labs/sfquickstarts/actions/runs/18788729669/workflow"                                        
#>   [8] "https://github.com/Snowflake-Labs/sfquickstarts/actions/runs/18790526152"                                                 
#>   [9] "https://github.com/Snowflake-Labs/sfquickstarts/actions/runs/18790526152/workflow"                                        
#>  [10] "https://github.com/Snowflake-Labs/sfquickstarts/actions/runs/18790943610"                                                 
#>  [11] "https://github.com/Snowflake-Labs/sfquickstarts/actions/runs/18790943610/workflow"                                        
#>  [12] "https://github.com/Snowflake-Labs/sfquickstarts/actions/workflows/validate-and-stage.yml"                                 
#>  [13] "https://github.com/Snowflake-Labs/sfquickstarts/activity"                                                                 
#>  [14] "https://github.com/Snowflake-Labs/sfquickstarts/blob/master/.Rhistory"                                                    
#>  [15] "https://github.com/Snowflake-Labs/sfquickstarts/blob/master/.gitignore"                                                   
#>  [16] "https://github.com/Snowflake-Labs/sfquickstarts/blob/master/.nvmrc"                                                       
#>  [17] "https://github.com/Snowflake-Labs/sfquickstarts/blob/master/.travis.yml"                                                  
#>  [18] "https://github.com/Snowflake-Labs/sfquickstarts/blob/master/BUILD.bazel"                                                  
#>  [19] "https://github.com/Snowflake-Labs/sfquickstarts/blob/master/FORMAT-GUIDE.md"                                              
#>  [20] "https://github.com/Snowflake-Labs/sfquickstarts/blob/master/LEGAL.md"                                                     
#>  [21] "https://github.com/Snowflake-Labs/sfquickstarts/blob/master/LICENSE"                                                      
#>  [22] "https://github.com/Snowflake-Labs/sfquickstarts/blob/master/MODULE.bazel"                                                 
#>  [23] "https://github.com/Snowflake-Labs/sfquickstarts/blob/master/MODULE.bazel.lock"                                            
#>  [24] "https://github.com/Snowflake-Labs/sfquickstarts/blob/master/README.md"                                                    
#>  [25] "https://github.com/Snowflake-Labs/sfquickstarts/blob/master/SECURITY.md"                                                  
#>  [26] "https://github.com/Snowflake-Labs/sfquickstarts/blob/master/WORKSPACE"                                                    
#>  [27] "https://github.com/Snowflake-Labs/sfquickstarts/blob/master/package.json"                                                 
#>  [28] "https://github.com/Snowflake-Labs/sfquickstarts/blob/master/sfquickstarts.Rproj"                                          
#>  [29] "https://github.com/Snowflake-Labs/sfquickstarts/blob/master/site/app/styles/_overrides.scss"                              
#>  [30] "https://github.com/Snowflake-Labs/sfquickstarts/branches"                                                                 
#>  [31] "https://github.com/Snowflake-Labs/sfquickstarts/commits/master"                                                           
#>  [32] "https://github.com/Snowflake-Labs/sfquickstarts/community"                                                                
#>  [33] "https://github.com/Snowflake-Labs/sfquickstarts/compare"                                                                  
#>  [34] "https://github.com/Snowflake-Labs/sfquickstarts/custom-properties"                                                        
#>  [35] "https://github.com/Snowflake-Labs/sfquickstarts/deployments"                                                              
#>  [36] "https://github.com/Snowflake-Labs/sfquickstarts/forks"                                                                    
#>  [37] "https://github.com/Snowflake-Labs/sfquickstarts/graphs/code-frequency"                                                    
#>  [38] "https://github.com/Snowflake-Labs/sfquickstarts/graphs/commit-activity"                                                   
#>  [39] "https://github.com/Snowflake-Labs/sfquickstarts/graphs/contributors"                                                      
#>  [40] "https://github.com/Snowflake-Labs/sfquickstarts/issues"                                                                   
#>  [41] "https://github.com/Snowflake-Labs/sfquickstarts/issues/2401"                                                              
#>  [42] "https://github.com/Snowflake-Labs/sfquickstarts/issues/2402"                                                              
#>  [43] "https://github.com/Snowflake-Labs/sfquickstarts/issues/2404"                                                              
#>  [44] "https://github.com/Snowflake-Labs/sfquickstarts/issues/2409"                                                              
#>  [45] "https://github.com/Snowflake-Labs/sfquickstarts/issues/2422"                                                              
#>  [46] "https://github.com/Snowflake-Labs/sfquickstarts/issues/2430/linked_closing_reference?reference_location=REPO_ISSUES_INDEX"
#>  [47] "https://github.com/Snowflake-Labs/sfquickstarts/issues/2437"                                                              
#>  [48] "https://github.com/Snowflake-Labs/sfquickstarts/issues/2448"                                                              
#>  [49] "https://github.com/Snowflake-Labs/sfquickstarts/issues/2449"                                                              
#>  [50] "https://github.com/Snowflake-Labs/sfquickstarts/issues/2466"                                                              
#>  [51] "https://github.com/Snowflake-Labs/sfquickstarts/issues/2471"                                                              
#>  [52] "https://github.com/Snowflake-Labs/sfquickstarts/issues/2473"                                                              
#>  [53] "https://github.com/Snowflake-Labs/sfquickstarts/issues/2481"                                                              
#>  [54] "https://github.com/Snowflake-Labs/sfquickstarts/issues/new/choose"                                                        
#>  [55] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Aissue%20state%3Aopen%20author%3AKevols"                     
#>  [56] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Aissue%20state%3Aopen%20author%3AShilpV"                     
#>  [57] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Aissue%20state%3Aopen%20author%3Acmareco"                    
#>  [58] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Aissue%20state%3Aopen%20author%3Aharrison-burchfield"        
#>  [59] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Aissue%20state%3Aopen%20author%3Asauravcal"                  
#>  [60] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Aissue%20state%3Aopen%20author%3Asfc-gh-anowlan"             
#>  [61] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Aissue%20state%3Aopen%20author%3Asfc-gh-dchaffelson"         
#>  [62] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Aissue%20state%3Aopen%20author%3Asfc-gh-mpenumathsa"         
#>  [63] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Aissue%20state%3Aopen%20author%3Asfc-gh-sweingartner"        
#>  [64] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Aissue%20state%3Aopen%20author%3Asfc-gh-wzou"                
#>  [65] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Aissue%20state%3Aopen%20author%3Asimakowskiw"                
#>  [66] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Apr+is%3Aopen+author%3AEveerasingam"                         
#>  [67] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Apr+is%3Aopen+author%3AG-r-ay"                               
#>  [68] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Apr+is%3Aopen+author%3Aadieado"                              
#>  [69] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Apr+is%3Aopen+author%3Aannettechen"                          
#>  [70] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Apr+is%3Aopen+author%3Acorydonbaylor"                        
#>  [71] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Apr+is%3Aopen+author%3Adgnestsn"                             
#>  [72] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Apr+is%3Aopen+author%3Afjaguero"                             
#>  [73] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Apr+is%3Aopen+author%3Ajamescha-earley"                      
#>  [74] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Apr+is%3Aopen+author%3Ajunkai-roe-ai"                        
#>  [75] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Apr+is%3Aopen+author%3Akanwalzs"                             
#>  [76] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Apr+is%3Aopen+author%3Amando222"                             
#>  [77] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Apr+is%3Aopen+author%3Aquintonwall"                          
#>  [78] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Apr+is%3Aopen+author%3Asfc-gh-apaniaguaguerrero"             
#>  [79] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Apr+is%3Aopen+author%3Asfc-gh-chammond"                      
#>  [80] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Apr+is%3Aopen+author%3Asfc-gh-etolotti"                      
#>  [81] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Apr+is%3Aopen+author%3Asfc-gh-hgallagher"                    
#>  [82] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Apr+is%3Aopen+author%3Asfc-gh-jreini"                        
#>  [83] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Apr+is%3Aopen+author%3Asfc-gh-knjeru"                        
#>  [84] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Apr+is%3Aopen+author%3Asfc-gh-mfus"                          
#>  [85] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Apr+is%3Aopen+author%3Asfc-gh-mnixon"                        
#>  [86] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Apr+is%3Aopen+author%3Asfc-gh-twhite"                        
#>  [87] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Apr+is%3Aopen+author%3Ashamernick1"                          
#>  [88] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Apr+is%3Aopen+author%3Avinodhini-sd"                         
#>  [89] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Apr+is%3Aopen+author%3Azhengchaoyan"                         
#>  [90] "https://github.com/Snowflake-Labs/sfquickstarts/issues?q=is%3Apr+is%3Aopen+no%3Amilestone"                                
#>  [91] "https://github.com/Snowflake-Labs/sfquickstarts/labels"                                                                   
#>  [92] "https://github.com/Snowflake-Labs/sfquickstarts/milestones"                                                               
#>  [93] "https://github.com/Snowflake-Labs/sfquickstarts/network"                                                                  
#>  [94] "https://github.com/Snowflake-Labs/sfquickstarts/network/dependencies"                                                     
#>  [95] "https://github.com/Snowflake-Labs/sfquickstarts/projects?is_search=true&query=is%3Aclosed"                                
#>  [96] "https://github.com/Snowflake-Labs/sfquickstarts/projects?is_search=true&query=is%3Aopen"                                  
#>  [97] "https://github.com/Snowflake-Labs/sfquickstarts/projects?query=is%3Aopen"                                                 
#>  [98] "https://github.com/Snowflake-Labs/sfquickstarts/projects?query=is%3Aopen+is%3Atemplate"                                   
#>  [99] "https://github.com/Snowflake-Labs/sfquickstarts/pull/1867"                                                                
#> [100] "https://github.com/Snowflake-Labs/sfquickstarts/pull/1870"                                                                
#> [101] "https://github.com/Snowflake-Labs/sfquickstarts/pull/1900"                                                                
#> [102] "https://github.com/Snowflake-Labs/sfquickstarts/pull/1910"                                                                
#> [103] "https://github.com/Snowflake-Labs/sfquickstarts/pull/1948"                                                                
#> [104] "https://github.com/Snowflake-Labs/sfquickstarts/pull/1992"                                                                
#> [105] "https://github.com/Snowflake-Labs/sfquickstarts/pull/2012"                                                                
#> [106] "https://github.com/Snowflake-Labs/sfquickstarts/pull/2057"                                                                
#> [107] "https://github.com/Snowflake-Labs/sfquickstarts/pull/2079"                                                                
#> [108] "https://github.com/Snowflake-Labs/sfquickstarts/pull/2098"                                                                
#> [109] "https://github.com/Snowflake-Labs/sfquickstarts/pull/2115"                                                                
#> [110] "https://github.com/Snowflake-Labs/sfquickstarts/pull/2116"                                                                
#> [111] "https://github.com/Snowflake-Labs/sfquickstarts/pull/2145"                                                                
#> [112] "https://github.com/Snowflake-Labs/sfquickstarts/pull/2172"                                                                
#> [113] "https://github.com/Snowflake-Labs/sfquickstarts/pull/2196"                                                                
#> [114] "https://github.com/Snowflake-Labs/sfquickstarts/pull/2223"                                                                
#> [115] "https://github.com/Snowflake-Labs/sfquickstarts/pull/2272"                                                                
#> [116] "https://github.com/Snowflake-Labs/sfquickstarts/pull/2273"                                                                
#> [117] "https://github.com/Snowflake-Labs/sfquickstarts/pull/2281"                                                                
#> [118] "https://github.com/Snowflake-Labs/sfquickstarts/pull/2428"                                                                
#> [119] "https://github.com/Snowflake-Labs/sfquickstarts/pull/2430"                                                                
#> [120] "https://github.com/Snowflake-Labs/sfquickstarts/pull/2477"                                                                
#> [121] "https://github.com/Snowflake-Labs/sfquickstarts/pull/2479"                                                                
#> [122] "https://github.com/Snowflake-Labs/sfquickstarts/pull/2480"                                                                
#> [123] "https://github.com/Snowflake-Labs/sfquickstarts/pull/2491"                                                                
#> [124] "https://github.com/Snowflake-Labs/sfquickstarts/pull/2492"                                                                
#> [125] "https://github.com/Snowflake-Labs/sfquickstarts/pull/2493"                                                                
#> [126] "https://github.com/Snowflake-Labs/sfquickstarts/pulls"                                                                    
#> [127] "https://github.com/Snowflake-Labs/sfquickstarts/pulls?page=2&q=is%3Apr+is%3Aopen"                                         
#> [128] "https://github.com/Snowflake-Labs/sfquickstarts/pulls?page=3&q=is%3Apr+is%3Aopen"                                         
#> [129] "https://github.com/Snowflake-Labs/sfquickstarts/pulls?page=4&q=is%3Apr+is%3Aopen"                                         
#> [130] "https://github.com/Snowflake-Labs/sfquickstarts/pulls?q=is%3Aopen+is%3Apr"                                                
#> [131] "https://github.com/Snowflake-Labs/sfquickstarts/pulls?q=is%3Apr+is%3Aclosed"                                              
#> [132] "https://github.com/Snowflake-Labs/sfquickstarts/pulls?q=is%3Apr+is%3Aopen+no%3Aassignee"                                  
#> [133] "https://github.com/Snowflake-Labs/sfquickstarts/pulls?q=is%3Apr+is%3Aopen+review%3Aapproved"                              
#> [134] "https://github.com/Snowflake-Labs/sfquickstarts/pulls?q=is%3Apr+is%3Aopen+review%3Achanges-requested"                     
#> [135] "https://github.com/Snowflake-Labs/sfquickstarts/pulls?q=is%3Apr+is%3Aopen+review%3Anone"                                  
#> [136] "https://github.com/Snowflake-Labs/sfquickstarts/pulls?q=is%3Apr+is%3Aopen+review%3Arequired"                              
#> [137] "https://github.com/Snowflake-Labs/sfquickstarts/pulls?q=is%3Apr+is%3Aopen+sort%3Acomments-asc"                            
#> [138] "https://github.com/Snowflake-Labs/sfquickstarts/pulls?q=is%3Apr+is%3Aopen+sort%3Acomments-desc"                           
#> [139] "https://github.com/Snowflake-Labs/sfquickstarts/pulls?q=is%3Apr+is%3Aopen+sort%3Acreated-asc"                             
#> [140] "https://github.com/Snowflake-Labs/sfquickstarts/pulls?q=is%3Apr+is%3Aopen+sort%3Areactions-%2B1-desc"                     
#> [141] "https://github.com/Snowflake-Labs/sfquickstarts/pulls?q=is%3Apr+is%3Aopen+sort%3Areactions--1-desc"                       
#> [142] "https://github.com/Snowflake-Labs/sfquickstarts/pulls?q=is%3Apr+is%3Aopen+sort%3Areactions-eyes-desc"                     
#> [143] "https://github.com/Snowflake-Labs/sfquickstarts/pulls?q=is%3Apr+is%3Aopen+sort%3Areactions-heart-desc"                    
#> [144] "https://github.com/Snowflake-Labs/sfquickstarts/pulls?q=is%3Apr+is%3Aopen+sort%3Areactions-rocket-desc"                   
#> [145] "https://github.com/Snowflake-Labs/sfquickstarts/pulls?q=is%3Apr+is%3Aopen+sort%3Areactions-smile-desc"                    
#> [146] "https://github.com/Snowflake-Labs/sfquickstarts/pulls?q=is%3Apr+is%3Aopen+sort%3Areactions-tada-desc"                     
#> [147] "https://github.com/Snowflake-Labs/sfquickstarts/pulls?q=is%3Apr+is%3Aopen+sort%3Areactions-thinking_face-desc"            
#> [148] "https://github.com/Snowflake-Labs/sfquickstarts/pulls?q=is%3Apr+is%3Aopen+sort%3Arelevance-desc"                          
#> [149] "https://github.com/Snowflake-Labs/sfquickstarts/pulls?q=is%3Apr+is%3Aopen+sort%3Aupdated-asc"                             
#> [150] "https://github.com/Snowflake-Labs/sfquickstarts/pulls?q=is%3Apr+is%3Aopen+sort%3Aupdated-desc"                            
#> [151] "https://github.com/Snowflake-Labs/sfquickstarts/pulse"                                                                    
#> [152] "https://github.com/Snowflake-Labs/sfquickstarts/search?l=go"                                                              
#> [153] "https://github.com/Snowflake-Labs/sfquickstarts/search?l=html"                                                            
#> [154] "https://github.com/Snowflake-Labs/sfquickstarts/search?l=javascript"                                                      
#> [155] "https://github.com/Snowflake-Labs/sfquickstarts/search?l=jupyter-notebook"                                                
#> [156] "https://github.com/Snowflake-Labs/sfquickstarts/search?l=python"                                                          
#> [157] "https://github.com/Snowflake-Labs/sfquickstarts/search?l=scss"                                                            
#> [158] "https://github.com/Snowflake-Labs/sfquickstarts/security"                                                                 
#> [159] "https://github.com/Snowflake-Labs/sfquickstarts/stargazers"                                                               
#> [160] "https://github.com/Snowflake-Labs/sfquickstarts/tags"                                                                     
#> [161] "https://github.com/Snowflake-Labs/sfquickstarts/tree/master/.github/ISSUE_TEMPLATE"                                       
#> [162] "https://github.com/Snowflake-Labs/sfquickstarts/tree/master/.vscode"                                                      
#> [163] "https://github.com/Snowflake-Labs/sfquickstarts/tree/master/claat"                                                        
#> [164] "https://github.com/Snowflake-Labs/sfquickstarts/tree/master/codelab-elements"                                             
#> [165] "https://github.com/Snowflake-Labs/sfquickstarts/tree/master/site"                                                         
#> [166] "https://github.com/Snowflake-Labs/sfquickstarts/tree/master/third_party"                                                  
#> [167] "https://github.com/Snowflake-Labs/sfquickstarts/tree/master/tmp/your-first-pwapp"                                         
#> [168] "https://github.com/Snowflake-Labs/sfquickstarts/tree/refs/heads/test/test-new-workflow-1"                                 
#> [169] "https://github.com/Snowflake-Labs/sfquickstarts/tree/refs/heads/test/test-new-workflow-2"                                 
#> [170] "https://github.com/Snowflake-Labs/sfquickstarts/watchers"                                                                 
# }