Convert files to markdown

Convert files to markdown

Usage

read_as_markdown(x, ..., canonical = FALSE)

Arguments

x

A filepath or url

...

These dots are for future extensions and must be empty.

canonical

logical, whether to postprocess the output from MarkItDown with commonmark::markdown_commonmark().

Value

A single string of markdown

Examples

# convert html
read_as_markdown("https://r4ds.hadley.nz/base-R.html") |>
  substr(1, 1000) |> cat()
#> 1. [Program](./program.html)
#> 2. [27  A field guide to base R](./base-R.html)
#> 
#> [R for Data Science (2e)](./)
#> 
#> * [Welcome](./index.html)
#> * [Preface to the second edition](./preface-2e.html)
#> * [Introduction](./intro.html)
#> * [Whole game](./whole-game.html)
#> 
#>   + [1  Data visualization](./data-visualize.html)
#>   + [2  Workflow: basics](./workflow-basics.html)
#>   + [3  Data transformation](./data-transform.html)
#>   + [4  Workflow: code style](./workflow-style.html)
#>   + [5  Data tidying](./data-tidy.html)
#>   + [6  Workflow: scripts and projects](./workflow-scripts.html)
#>   + [7  Data import](./data-import.html)
#>   + [8  Workflow: getting help](./workflow-help.html)
#> * [Visualize](./visualize.html)
#> 
#>   + [9  Layers](./layers.html)
#>   + [10  Exploratory data analysis](./EDA.html)
#>   + [11  Communication](./communication.html)
#> * [Transform](./transform.html)
#> 
#>   + [12  Logical vectors](./logicals.html)
#>   + [13  Numbers](./numbers.html)
#>   + [14  Strings](./strings.html)
#>   + [15  Regular expressions](./regexp

read_as_markdown("https://r4ds.hadley.nz/base-R.html", canonical = TRUE) |>
  substr(1, 1000) |> cat()
#> 1.  [Program](./program.html)
#> 2.  [27  A field guide to base R](./base-R.html)
#> 
#> [R for Data Science (2e)](./)
#> 
#>   - [Welcome](./index.html)
#> 
#>   - [Preface to the second edition](./preface-2e.html)
#> 
#>   - [Introduction](./intro.html)
#> 
#>   - [Whole game](./whole-game.html)
#>     
#>       - [1  Data visualization](./data-visualize.html)
#>       - [2  Workflow: basics](./workflow-basics.html)
#>       - [3  Data transformation](./data-transform.html)
#>       - [4  Workflow: code style](./workflow-style.html)
#>       - [5  Data tidying](./data-tidy.html)
#>       - [6  Workflow: scripts and projects](./workflow-scripts.html)
#>       - [7  Data import](./data-import.html)
#>       - [8  Workflow: getting help](./workflow-help.html)
#> 
#>   - [Visualize](./visualize.html)
#>     
#>       - [9  Layers](./layers.html)
#>       - [10  Exploratory data analysis](./EDA.html)
#>       - [11  Communication](./communication.html)
#> 
#>   - [Transform](./transform.html)
#>     
#>       - [12  Logical vectors](./logicals.html)
#>       - [13  Numbers](./num

# convert pdf
pdf <- file.path(R.home("doc"), "NEWS.pdf")
read_as_markdown(pdf) |> substr(1, 1000) |> cat()
#> NEWS for R version 4.4.2 (2024-10-31)
#> 
#> NEWS
#> 
#> R News
#> 
#> CHANGES IN R 4.4.2
#> 
#> C-LEVEL FACILITIES:
#> 
#> (cid:136) The S-compatibility macros F77_COM and F77_COMDECL de(cid:28)ned in header
#> ‘R_ext/RS.h’ are deprecated and will be removed shortly. We could (cid:28)nd no record of
#> their use.
#> 
#> BUG FIXES:
#> 
#> (cid:136) Mathlib function lgammacor(x) no longer warns about under(cid:29)ow to zero for large x.
#> (cid:136) Text widths and heights were incorrectly reported by the Quartz device if the drawing
#> context didn’t exist yet (typically when drawing o(cid:27)-screen to a window that is yet to
#> appear, see PR#18591).
#> 
#> (cid:136) The Quartz device could segfault in cases where paths with spaces are used in the
#> 
#> new glyph drawing API. Thanks to Tomek Gieorgijewski (PR#18758).
#> 
#> (cid:136) On macOS in R CRAN builds, it is again possible to read little-endian UTF-16 text
#> with a BOM from a connection using encoding="UTF-16". Users building R from
#> source should avoid using the system libiconv in macOS 14.1 and late
## alternative:
# pdftools::pdf_text(pdf) |> substr(1, 2000) |> cat()